Reorganize test rulses under //rules package. Add compose unit test support
