+++ rules/android/BUILD.bazel
@@ -1,0 +1,3 @@
+load("@grab_bazel_common//tools/android_mock:andorid_mock.bzl", "mock_android_jar")
+
+mock_android_jar()
