+++ rules/android/runtime_resources.bzl
