+++ tools/android_mock/BUILD.bazel
@@ -1,0 +1,27 @@
+load("@io_bazel_rules_kotlin//kotlin:jvm.bzl", "kt_jvm_library", "kt_jvm_test")
+
+kt_jvm_library(
+    name = "mocked_android_jar_lib",
+    srcs = glob([
+        "src/main/java/**/*.kt",
+    ]),
+    visibility = [
+        "//visibility:private",
+    ],
+    deps = [
+        "@bazel_common_maven//:com_github_ajalt_clikt",
+        "@bazel_common_maven//:org_ow2_asm_asm",
+        "@bazel_common_maven//:org_ow2_asm_asm_tree",
+    ],
+)
+
+java_binary(
+    name = "mocked_android_jar_generator",
+    main_class = "com.grab.test.MainKt",
+    visibility = [
+        "//visibility:public",
+    ],
+    runtime_deps = [
+        ":mocked_android_jar_lib",
+    ],
+)
