--- tools/kotlin/BUILD
@@ -1,1 +1,0 @@
-load(":android.bzl", "kt_android_library")
