--- tools/test/BUILD.bazel
@@ -1,3 +1,0 @@
-load("@grab_bazel_common//tools/test:mockable_android_jar.bzl", "mockable_android_jar")
-
-mockable_android_jar()
