--- tools/test/android/BUILD.bazel
@@ -1,27 +1,0 @@
-load("@grab_bazel_common//tools/kotlin:android.bzl", "kt_android_library")
-load("@grab_bazel_common//tools/test:test.bzl", "grab_android_local_test")
-
-kt_android_library(
-    name = "grab_android_local_test_lib",
-    srcs = glob([
-        "src/main/java/**/*.kt",
-    ]),
-    deps = [
-        "@bazel_common_maven//:com_google_android_material_material",
-    ],
-)
-
-grab_android_local_test(
-    name = "grab_android_local_test",
-    srcs = glob([
-        "src/test/java/**/*.kt",
-    ]),
-    associates = [
-        ":grab_android_local_test_lib_kt",
-    ],
-    deps = [
-        ":grab_android_local_test_lib",
-        "@maven//:junit_junit",
-        "@maven//:org_json_json",
-    ],
-)
