--- tools/test/jvm/BUILD.bazel
@@ -1,22 +1,0 @@
-load("@grab_bazel_common//tools/test:test.bzl", "grab_kt_jvm_test")
-load("@io_bazel_rules_kotlin//kotlin:jvm.bzl", "kt_jvm_library")
-
-kt_jvm_library(
-    name = "grab_kt_jvm",
-    srcs = glob([
-        "src/main/java/**/*.kt",
-    ]),
-)
-
-grab_kt_jvm_test(
-    name = "grab_kt_jvm_test",
-    srcs = glob([
-        "src/test/java/**/*.kt",
-    ]),
-    associates = [
-        ":grab_kt_jvm",
-    ],
-    deps = [
-        "@maven//:junit_junit",
-    ],
-)
