+++ tools/worker/BUILD.bazel
@@ -1,5 +1,5 @@
 load("@io_bazel_rules_kotlin//kotlin:jvm.bzl", "kt_jvm_library")
-load("@grab_bazel_common//tools/test:test.bzl", "grab_kt_jvm_test")
+load("@grab_bazel_common//rules:defs.bzl", "kotlin_test")
 
 kt_jvm_library(
     name = "worker_lib",
@@ -20,7 +20,7 @@ kt_jvm_library(
     ],
 )
 
-grab_kt_jvm_test(
+kotlin_test(
     name = "worker_test",
     srcs = glob([
         "src/test/kotlin/**/*.kt",
